Implement DynamicBuffer_v2 (API Change):
Boost.Asio made a change to the DynamicBuffer concept. The
old concept is called DynamicBuffer_v1 and the new one is
called DynamicBuffer_v2. To unify the Beast and Asio
implementations, these breaking changes are made:

* The concept in Beast is now called DynamicStorage, and
  represents a true storage type.

* The DynamicStorage concept requirements are updated to
  include DynamicBuffer_v2

* All Beast models of DynamicStorage implement DynamicBuffer_v2

* The function beast::dynamic_buffer() wraps a DynamicStorage
  object in a copyable wrapper which meets the requirements
  of DynamicBuffer_v2

* All Beast algorithms work with DynamicBuffer_v2

The consequence of this change is that Beast and Asio buffer
types and algorithms may now interoperate as expected.